Stop using CSS classes for element lookup
On first implementing this site ca 2014, I went for an absolutist
twist on progressive enhancement wherein I wanted it to be
impossible to tell by examining the HTML source whether it was
designed to be manipulated by JavaScript or not. This never made
sense except as an exercise. This commit makes the code which does
DOM queries that aren't for updating CSS styling use attribute
values instead. This makes it clearer what's there for presentation
and what's there for to hook into JS logic.